+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v1.2.2
 * adding diffs
+* changing raw read to raw_get 
+* adding raw_post
 
 ## v1.2.1
 * crosswalk endpoint change